requirements.txt
torch
torchvision
numpy
pillow
scikit-learn
scipy
matplotlib
tqdm
accelerate
